[qa] Aligned formatting checks
Made QA checks more consistent with openwrt-openwisp-monitoring and openwisp-utils.

+++ openwisp-config/files/sbin/openwisp-update-config.lua
@@ -25,8 +25,7 @@ local check_dir = tmp_dir .. '/check'
 local check_config_dir = check_dir .. '/etc/config'
 local downloaded_conf = conf or (tmp_dir .. '/configuration.tar.gz')
 local openwisp_dir = not TEST and '/etc/openwisp' or working_dir .. '/openwisp'
-local standard_config_dir = not TEST and '/etc/config' or working_dir ..
-                              '/update-test/etc/config'
+local standard_config_dir = not TEST and '/etc/config' or working_dir .. '/update-test/etc/config'
 local test_root_dir = working_dir .. '/update-test'
 local remote_dir = openwisp_dir .. '/remote'
 local remote_config_dir = remote_dir .. '/etc/config'
@@ -34,7 +33,9 @@ local stored_dir = openwisp_dir .. '/stored'
 local stored_config_dir = stored_dir .. '/etc/config'
 local added_file = openwisp_dir .. '/added.list'
 local modified_file = openwisp_dir .. '/modified.list'
-local get_standard = function() return uci.cursor(standard_config_dir) end
+local get_standard = function()
+  return uci.cursor(standard_config_dir)
+end
 local get_remote = function()
   return uci.cursor(remote_config_dir, '/tmp/openwisp/.uci')
 end
@@ -119,17 +120,13 @@ if lfs.attributes(remote_config_dir, 'mode') == 'directory' then
           end
           -- remove entire section if empty
           local result = standard:get_all(file, section['.name'])
-          if result and utils.is_uci_empty(result) then
-            standard:delete(file, section['.name'])
-          end
+          if result and utils.is_uci_empty(result) then standard:delete(file, section['.name']) end
         end
       end
       standard:commit(file)
       -- remove uci file if empty
       local uci_file = standard:get_all(file)
-      if uci_file and utils.is_table_empty(uci_file) then
-        os.remove(standard_path)
-      end
+      if uci_file and utils.is_table_empty(uci_file) then os.remove(standard_path) end
     end
   end
 end
@@ -186,7 +183,9 @@ local modified_changed = false
 
 -- loop each file except directories and standard UCI config files
 local ignored_path = remote_config_dir .. '/'
-local function is_ignored(path) return utils.starts_with(path, ignored_path) end
+local function is_ignored(path)
+  return utils.starts_with(path, ignored_path)
+end
 
 for path, attr in utils.dirtree(remote_dir) do
   if attr.mode == 'file' and not is_ignored(path) then
@@ -223,8 +222,7 @@ for path, attr in utils.dirtree(remote_dir) do
       end
       -- add file to filesystem
       os.execute('mkdir -p ' .. utils.escape_shell_arg(dest_dir))
-      os.execute('cp ' .. utils.escape_shell_arg(path) .. ' ' ..
-                   utils.escape_shell_arg(dest_path))
+      os.execute('cp ' .. utils.escape_shell_arg(path) .. ' ' .. utils.escape_shell_arg(dest_path))
     end
   end
 end
